libfiber can be used on MacOS

+++ lib_fiber/samples/fiber_local/main.c
@@ -12,7 +12,7 @@ static __thread int __local_key;
 static void free_fn(void *ctx)
 {
 	printf("thread-%ld, fiber-%d: free buf\r\n",
-		acl_pthread_self(), acl_fiber_self());
+		(long) acl_pthread_self(), acl_fiber_self());
 	acl_myfree(ctx);
 }
 
@@ -34,7 +34,7 @@ static void fiber_main(ACL_FIBER *fiber acl_unused, void *ctx acl_unused)
 	}
 
 	printf("thread-%ld, fiber-%d: __local_key: %d, buf: %s\r\n",
-		acl_pthread_self(), acl_fiber_self(), __local_key, buf);
+		(long) acl_pthread_self(), acl_fiber_self(), __local_key, buf);
 	return;
 
 	if (--__left_fiber == 0) {
